refactor: use notion2hast library
