+++ package.json
@@ -48,6 +48,7 @@
     "instantsearch.css": "^8.3.0",
     "luxon": "^3.4.4",
     "mdast-util-to-string": "^4.0.0",
+    "notion2hast": "^0.1.3",
     "postcss": "^8.4.40",
     "postcss-import": "^16.1.0",
     "postcss-js": "^4.0.1",
